minor refacot
