+++ packages/backend/src/core/liveness/LivenessIndexer.ts
@@ -1,5 +1,5 @@
 import { assert, Logger } from '@l2beat/backend-tools'
-import { Hash256, notUndefined, UnixTime } from '@l2beat/shared-pure'
+import { Hash256, UnixTime } from '@l2beat/shared-pure'
 import { ChildIndexer } from '@l2beat/uif'
 import { Knex } from 'knex'
 
@@ -8,19 +8,15 @@ import {
   IndexerStateRecord,
   IndexerStateRepository,
 } from '../../peripherals/database/IndexerStateRepository'
-import {
-  LivenessConfigurationRecord,
-  LivenessConfigurationRepository,
-} from '../../peripherals/database/LivenessConfigurationRepository'
+import { LivenessConfigurationRepository } from '../../peripherals/database/LivenessConfigurationRepository'
 import { LivenessRepository } from '../../peripherals/database/LivenessRepository'
 import { HourlyIndexer } from './HourlyIndexer'
 import { LivenessClient } from './LivenessClient'
-import { LivenessFunctionCall, LivenessTransfer } from './types/LivenessConfig'
-import { LivenessConfigurationIdentifier } from './types/LivenessConfigurationIdentifier'
 import {
   adjustToForBigqueryCall,
   getLivenessConfigHash,
   isTimestampInRange,
+  mergeConfigs,
 } from './utils'
 import { processLivenessConfigurations } from './utils/processLivenessConfigurations'
 
@@ -51,7 +47,6 @@ export class LivenessIndexer extends ChildIndexer {
   override async update(from: number, to: number): Promise<number> {
     const { transfersConfig, functionCallsConfig, adjustedTo } =
       await this.getConfiguration(from, to)
-    //
 
     const data = await this.livenessClient.getLivenessData(
       transfersConfig,
@@ -73,6 +68,11 @@ export class LivenessIndexer extends ChildIndexer {
         ),
       )
     })
+
+    const usedConfigurationsIds = [
+      ...transfersConfig.map((c) => c.livenessConfigurationId),
+      ...functionCallsConfig.map((c) => c.livenessConfigurationId),
+    ]
 
     this.logger.info('Updated liveness data', {
       from,
@@ -234,48 +234,5 @@ export class LivenessIndexer extends ChildIndexer {
   // and the data will not be fetched again
   override async invalidate(targetHeight: number): Promise<number> {
     return Promise.resolve(targetHeight)
-  }
-}
-
-// TODO: move to separate file
-export function mergeConfigs(
-  projects: Project[],
-  configs: LivenessConfigurationRecord[],
-): {
-  transfers: LivenessTransfer[]
-  functionCalls: LivenessFunctionCall[]
-} {
-  // add proper values from configs
-  return {
-    transfers: projects
-      .flatMap((p) => p.livenessConfig?.transfers)
-      .filter(notUndefined)
-      .map((t) => {
-        const config = configs.find(
-          (c) => c.identifier === LivenessConfigurationIdentifier(t),
-        )
-        assert(config, 'Config should not be undefined there')
-
-        return {
-          ...t,
-          latestSyncedTimestamp: config.lastSyncedTimestamp,
-          livenessConfigurationId: config.id,
-        }
-      }),
-    functionCalls: projects
-      .flatMap((p) => p.livenessConfig?.functionCalls)
-      .filter(notUndefined)
-      .map((t) => {
-        const config = configs.find(
-          (c) => c.identifier === LivenessConfigurationIdentifier(t),
-        )
-        assert(config, 'Config should not be undefined there')
-
-        return {
-          ...t,
-          latestSyncedTimestamp: config.lastSyncedTimestamp,
-          livenessConfigurationId: config.id,
-        }
-      }),
   }
 }
